pyproject.toml
[project]
name = "lumi-real"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
